+++ packages/apollo-react-relay-duct-tape/src/storeObservation/__tests__/__generated__/compiledHooks_QueryTypeWatchNodeQuery.graphql.ts
@@ -102,3 +102,5 @@ return {
   }
 };
 })();
+
+export default documents;