+++ packages/graphql-codegen-typescript-operations/src/config.ts
@@ -120,4 +120,5 @@ export interface TypeScriptDocumentsPluginConfig extends RawDocumentsConfig {
    */
   addOperationExport?: boolean;
   baseTypesPath?: string;
+  legacyEnumsCompatibility?: boolean;
 }