+++ packages/ts-codegen/src/context/utilities.ts
@@ -8,7 +8,6 @@ export function createImportDeclaration(
 ) {
   return factory.createImportDeclaration(
     undefined,
-    undefined,
     factory.createImportClause(
       true,
       undefined,